Added RAM controller
Added a component called ram_controller, which allows the RAM to be read
or written depending on the inputs.
RAM address selection input has been added to the top architecture. Now
it is possible to read any data from the RAM from outside of the design.

However, WRITING HAS PRIORITY.